Removed a lot of redndant cast after generisizing, but many remain.

+++ javatools/src/main/java/org/xvm/asm/constants/BFloat16Constant.java
@@ -70,7 +70,7 @@ public class BFloat16Constant
      */
     @Override
     public Float getValue() {
-        return Float.valueOf(m_flVal);
+        return m_flVal;
     }
